requirements.txt
django==5.2
python-dotenv==0.20.0
djangorestframework==3.13.1
 # psycopg2-binary==2.9.3
# for PostgreSQL (if you're using it)
